+++ test/helper.rb
@@ -7,6 +7,7 @@ $LOAD_PATH.unshift(ROOT_PATH)
 
 # require pry for debugging (`binding.pry`)
 require 'pry'
+require 'test/support/factory'
 
 # Force tests to run without halting on failures (needed so all tests will run
 # properly).  For halt on fail behavior testing, the context of those tests
@@ -54,50 +55,7 @@ class TestContext < Assert::Context
   def self.method_added(meth)
     if meth.to_s =~ Assert::Suite::TEST_METHOD_REGEX
       ci = Assert::Suite::ContextInfo.new(self, Factory.context_info_called_from)
-      TEST_ASSERT_SUITE.tests << Assert::Test.new(meth.to_s, ci, meth)
+      TEST_ASSERT_SUITE.tests << Assert::Test.new(meth.to_s, ci, :code => meth)
     end
   end
-end
-
-module Factory
-
-  def self.context_info_called_from
-    "/path/to_file.rb:1234"
-  end
-
-  def self.context_info(context_class)
-    Assert::Suite::ContextInfo.new(context_class, context_info_called_from)
-  end
-
-  # Generate an anonymous `Context` inherited from `TestContext` by default.
-  # This provides a common interface for all contexts used in testing.
-
-  def self.context_class(inherit_from=nil, &block)
-    inherit_from ||= TestContext
-    klass = Class.new(inherit_from, &block)
-    default = (const_name = "FactoryAssertContext").dup
-
-    while(Object.const_defined?(const_name)) do
-      const_name = "FactoryAssertContext#{rand(Time.now.to_i)}"
-    end
-    Object.const_set(const_name, klass)
-    klass
-  end
-
-  # Generate a no-op test for use in testing.
-
-  def self.test(*args, &block)
-    name = (args[0] || "a test").to_s
-    context_info = args[1] || self.context_info(self.context_class)
-    test_block = (block || args[2] || ::Proc.new{})
-
-    Assert::Test.new(name, context_info, &test_block)
-  end
-
-  # Generate a skip result for use in testing.
-
-  def self.skip_result(name, exception)
-    Assert::Result::Skip.new(Factory.test(name), exception)
-  end
-
 end
