+++ test/unit/test_tests.rb
@@ -9,7 +9,7 @@ class Assert::Test
       @test_code = lambda{ assert(true) }
       @context_class = Factory.context_class{ desc "context class" }
       @context_info  = Factory.context_info(@context_class)
-      @test = Factory.test("should do something amazing", @context_info, @test_code)
+      @test = Factory.test("should do something amazing", @context_info, :code => @test_code)
     end
     teardown do
       TEST_ASSERT_SUITE.tests.clear
